http-api: changed eventsource tool
