+++ pkg/npm/http-api/example/node.js
@@ -1,17 +1,32 @@
 /*
  * ATTENTION: The "eval" devtool has been used (maybe by default in mode: "development").
- * This devtool is not neither made for production nor for readable output files.
+ * This devtool is neither made for production nor for readable output files.
  * It uses "eval()" calls to create a separate source file in the browser devtools.
  * If you are trying to read the output file, select a different devtool (https://webpack.js.org/configuration/devtool/)
  * or disable the default devtool with "devtool: false".
  * If you are looking for production-ready output files, see mode: "production" (https://webpack.js.org/configuration/mode/).
  */
 /******/ (() => { // webpackBootstrap
+/******/ 	var __webpack_modules__ = ({
+
+/***/ "./src/example/node.js":
 /*!*****************************!*\
   !*** ./src/example/node.js ***!
   \*****************************/
-/*! unknown exports (runtime-defined) */
-/*! runtime requirements:  */
+/***/ (() => {
+
 eval("// import Urbit from '../../dist/index';\n// async function blastOff() {\n//   const airlock = await Urbit.authenticate({\n//     ship: 'zod',\n//     url: 'localhost:8080',\n//     code: 'lidlut-tabwed-pillex-ridrup',\n//     verbose: true\n//   });\n//   airlock.subscribe('chat-view', '/primary');\n// }\n// blastOff();\n\n//# sourceURL=webpack://@urbit/http-api/./src/example/node.js?");
+
+/***/ })
+
+/******/ 	});
+/************************************************************************/
+/******/ 	
+/******/ 	// startup
+/******/ 	// Load entry module and return exports
+/******/ 	// This entry module can't be inlined because the eval devtool is used.
+/******/ 	var __webpack_exports__ = {};
+/******/ 	__webpack_modules__["./src/example/node.js"]();
+/******/ 	
 /******/ })()
 ;
