--- pkg/npm/http-api/index.js
@@ -1,2 +1,0 @@
-import Urbit from './dist';
-export { Urbit as default, Urbit };
