+++ pkg/npm/http-api/package.json
@@ -8,20 +8,15 @@
     "url": "ssh://git@github.com/urbit/urbit.git",
     "directory": "pkg/npm/http-api"
   },
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "browser": "dist/esm/index.js",
-  "types": "dist/esm/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist",
     "src"
   ],
-  "engines": {
-    "node": ">=13"
-  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "npm run clean && webpack --config webpack.prod.js && tsc -p tsconfig.json && tsc -p tsconfig-cjs.json",
+    "build": "npm run clean && tsc -p tsconfig.json",
     "clean": "rm -rf dist/*"
   },
   "peerDependencies": {},
@@ -38,31 +33,29 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
     "@babel/plugin-proposal-optional-chaining": "^7.12.1",
     "@babel/preset-typescript": "^7.12.1",
+    "@types/browser-or-node": "^1.2.0",
     "@types/eventsource": "^1.1.5",
     "@types/react": "^16.9.56",
     "@typescript-eslint/eslint-plugin": "^4.7.0",
     "@typescript-eslint/parser": "^4.7.0",
-    "@types/browser-or-node": "^1.2.0",
     "babel-loader": "^8.2.1",
     "clean-webpack-plugin": "^3.0.0",
     "tslib": "^2.0.3",
     "typescript": "^3.9.7",
+    "util": "^0.12.3",
     "webpack": "^5.4.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.12.5",
+    "@microsoft/fetch-event-source": "^2.0.0",
+    "@urbit/api": "file:../api",
     "browser-or-node": "^1.3.0",
     "browserify-zlib": "^0.2.0",
-    "buffer": "^5.7.1",
-    "encoding": "^0.1.13",
-    "eventsource": "^1.0.7",
+    "buffer": "^6.0.3",
     "node-fetch": "^2.6.1",
     "stream-browserify": "^3.0.0",
-    "stream-http": "^3.1.1",
-    "util": "^0.12.3",
-    "xmlhttprequest": "^1.8.0",
-    "xmlhttprequest-ssl": "^1.6.0"
+    "stream-http": "^3.1.1"
   }
 }
