+++ pkg/npm/http-api/tsconfig.json
@@ -1,18 +1,19 @@
 {
-  "include": ["src/**/*.ts"],
+  "include": ["src/*.ts"],
   "exclude": ["node_modules", "dist", "@types"],
   "compilerOptions": {
-	  "outDir": "./dist/esm",
-    "module": "ES2020",
+	  "outDir": "./dist",
+    "module": "ESNext",
     "noImplicitAny": true,
-    "target": "ES2020",
+    "target": "ESNext",
     "pretty": true,
     "moduleResolution": "node",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "declaration": true,
     "sourceMap": true,
-    "strict": false
-    // "lib": ["ES2020"],
+    "strict": false,
+    "noErrorTruncation": true,
+    "allowJs": true,
   }
 }
